backend/requirements.txt
Flask==2.3.3
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
nada-dsl==0.9.0
nada-numpy==0.9.0
nada-ai==0.9.0
nillion-client==0.35.0
gunicorn==21.2.0
Werkzeug==2.3.7